+++ geoopt/manifolds/sphere.py
@@ -2,7 +2,7 @@ import torch
 
 from .base import Manifold
 from ..tensor import ManifoldTensor
-from ..utils import size2shape
+from ..utils import size2shape, broadcast_shapes
 import geoopt.linalg.batch_linalg
 
 __all__ = ["Sphere", "SphereExact"]
@@ -41,7 +41,9 @@ class Sphere(Manifold):
     name = "Sphere"
     reversible = False
 
-    def __init__(self, intersection=None, complement=None):
+    def __init__(
+        self, intersection: torch.Tensor = None, complement: torch.Tensor = None
+    ):
         super().__init__()
         if intersection is not None and complement is not None:
             raise TypeError(
@@ -84,7 +86,7 @@ class Sphere(Manifold):
                 )
         return ok, reason
 
-    def _check_point_on_manifold(self, x, *, atol=1e-5, rtol=1e-5):
+    def _check_point_on_manifold(self, x: torch.Tensor, *, atol=1e-5, rtol=1e-5):
         norm = x.norm(dim=-1)
         ok = torch.allclose(norm, norm.new((1,)).fill_(1), atol=atol, rtol=rtol)
         if not ok:
@@ -99,66 +101,72 @@ class Sphere(Manifold):
             )
         return True, None
 
-    def _check_vector_on_tangent(self, x, u, *, atol=1e-5, rtol=1e-5):
-        inner = self.inner(None, x, u, keepdim=True)
+    def _check_vector_on_tangent(
+        self, x: torch.Tensor, u: torch.Tensor, *, atol=1e-5, rtol=1e-5
+    ):
+        inner = self.inner(x, x, u, keepdim=True)
         ok = torch.allclose(inner, inner.new_zeros((1,)), atol=atol, rtol=rtol)
         if not ok:
             return False, "`<x, u> != 0` with atol={}, rtol={}".format(atol, rtol)
         return True, None
 
-    def inner(self, x, u, v=None, *, keepdim=False):
+    def inner(
+        self, x: torch.Tensor, u: torch.Tensor, v: torch.Tensor = None, *, keepdim=False
+    ):
         if v is None:
             v = u
-        return (u * v).sum(-1, keepdim=keepdim)
+        inner = (u * v).sum(-1, keepdim=keepdim)
+        target_shape = broadcast_shapes(x.shape[:-1] + (1,) * keepdim, inner.shape)
+        return inner.expand(target_shape)
 
-    def projx(self, x):
+    def projx(self, x: torch.Tensor):
         x = self._project_on_subspace(x)
         return x / x.norm(dim=-1, keepdim=True)
 
-    def proju(self, x, u):
+    def proju(self, x: torch.Tensor, u: torch.Tensor):
         u = u - (x * u).sum(dim=-1, keepdim=True) * x
         return self._project_on_subspace(u)
 
-    def expmap(self, x, u):
+    def expmap(self, x: torch.Tensor, u: torch.Tensor):
         norm_u = u.norm(dim=-1, keepdim=True)
         exp = x * torch.cos(norm_u) + u * torch.sin(norm_u) / norm_u
         retr = self.projx(x + u)
         cond = norm_u > EPS[norm_u.dtype]
         return torch.where(cond, exp, retr)
 
-    def retr(self, x, u):
+    def retr(self, x: torch.Tensor, u: torch.Tensor):
         return self.projx(x + u)
 
-    def transp(self, x, y, v):
+    def transp(self, x: torch.Tensor, y: torch.Tensor, v: torch.Tensor):
         return self.proju(y, v)
 
-    def logmap(self, x, y):
+    def logmap(self, x: torch.Tensor, y: torch.Tensor):
         u = self.proju(x, y - x)
         dist = self.dist(x, y, keepdim=True)
         # If the two points are "far apart", correct the norm.
         cond = dist.gt(EPS[dist.dtype])
         return torch.where(cond, u * dist / u.norm(dim=-1, keepdim=True), u)
 
-    def dist(self, x, y, *, keepdim=False):
-        inner = self.inner(None, x, y, keepdim=keepdim).clamp(-0.9999, 0.9999)
+    def dist(self, x: torch.Tensor, y: torch.Tensor, *, keepdim=False):
+        inner = self.inner(x, x, y, keepdim=keepdim).clamp(-0.9999, 0.9999)
         return torch.acos(inner)
 
     egrad2rgrad = proju
 
-    def _configure_manifold_complement(self, complement):
+    def _configure_manifold_complement(self, complement: torch.Tensor):
         Q, _ = geoopt.linalg.batch_linalg.qr(complement)
         P = -Q @ Q.transpose(-1, -2)
         P[..., torch.arange(P.shape[-2]), torch.arange(P.shape[-2])] += 1
         self.register_buffer("projector", P)
 
-    def _configure_manifold_intersection(self, intersection):
+    def _configure_manifold_intersection(self, intersection: torch.Tensor):
         Q, _ = geoopt.linalg.batch_linalg.qr(intersection)
         self.register_buffer("projector", Q @ Q.transpose(-1, -2))
 
     def _configure_manifold_no_constraints(self):
         self.register_buffer("projector", None)
 
-    def _project_on_subspace(self, x):
+    def _project_on_subspace(self, x: torch.Tensor):
         if self.projector is not None:
             return x @ self.projector.transpose(-1, -2)
         else:
@@ -203,6 +211,8 @@ class Sphere(Manifold):
                 *size, device=self.projector.device, dtype=self.projector.dtype
             )
         return ManifoldTensor(self.projx(tens), manifold=self)
+
+    random = random_uniform
 
 
 class SphereExact(Sphere):